pyproject.toml
[project]
name = "novelai-sdk"
version = "0.0.0"
description = "Python client library for NovelAI API with Pydantic validation"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "caru-ini", email = "caru17122@gmail.com" }
]
requires-python = ">=3.13"
keywords = ["novelai", "api", "image-generation", "ai", "stable-diffusion", "pydantic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.28.1",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
build = ["uv>=0.6.14"]

[project.urls]
Homepage = "https://github.com/caru-ini/novelai-api"
Repository = "https://github.com/caru-ini/novelai-api"
Issues = "https://github.com/caru-ini/novelai-api/issues"
Documentation = "https://github.com/caru-ini/novelai-api#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/novelai"]

[tool.poe.tasks]
fmt = "ruff format"
lint = "ruff check"
lint-fix = "ruff check --fix"
check = "pyright"
test = "pytest"
test-cov = "pytest --cov=src/novelai"
pre-commit = ["fmt", "lint", "check"]
release = "semantic-release publish"

[tool.semantic_release]
version_variables = ["src/novelai/__init__.py:__version__"]
branch = "main"
build_command = """
    python -m pip install -e '.[build]'
    python -m uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    python -m uv build
"""
dist_glob_patterns = ["dist/*"]
tag_format = "v{version}"

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "github-actions <actions@github.com>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "docs", "style", "refactor", "perf", "test", "chore", "build", "ci"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[dependency-groups]
dev = [
    "opencv-python>=4.10.0",
    "poethepoet>=0.37.0",
    "pre-commit>=4.0.1",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "python-semantic-release>=9.14.0",
    "ruff>=0.14.1",
]

